Make Snatch depend on generalized SnatchBase module.
