+++ Sources/Snatch/Result+Json.swift
@@ -1,14 +1,8 @@
 import Promise
+import SnatchBase
 import Foundation
 
 public extension Result {
-    /**
-        This error gets thrown in rejection to an attempt to extract json from a result with empty body data.
-    */
-    public class NoBodyError: Error {
-        var localizedDescription: String = "The response body is empty."
-    }
-
     /**
         Creates a promise that will attepmt to interpret the data of result AKA the body of http request as `type` TargetType type. If the body is empty, WILL REJECT.
 
@@ -18,7 +12,7 @@ public extension Result {
     */
     public func json<TargetType: Decodable>(_ type: TargetType.Type) -> Promise<TargetType> {
         guard let data = data else {
-            return Promise(error: NoBodyError())
+            return Promise(error: NoBodyError.new)
         }
 
         return Promise { fulfill, reject in
