--- Sources/Snatch/Util/URL+components.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-extension URL {
-    var components: NSURLComponents? {
-        return NSURLComponents(url: self, resolvingAgainstBaseURL: false)
-    }
-}
